extended fab in compose
