--- app/src/main/res/layout/content_main.xml
@@ -1,39 +1,0 @@
-<layout xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="com.mathewsachin.fategrandautomata.ui.prefs.MainSettingsViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:showIn="@layout/activity_main">
-
-        <androidx.fragment.app.FragmentContainerView
-            android:name="com.mathewsachin.fategrandautomata.ui.prefs.MainSettingsFragment"
-            android:id="@+id/main_pref"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-        <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-            android:theme="@style/Theme.MaterialComponents"
-            android:id="@+id/service_toggle_btn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="32dp"
-            app:icon="@{vm.serviceStarted ? @drawable/ic_close : @drawable/ic_launch}"
-            app:iconTint="@android:color/white"
-            android:backgroundTint="@{vm.serviceStarted ? @color/colorStopService : @color/colorPrimary}"
-            android:textColor="@android:color/white"
-            android:text="@{vm.serviceStarted ? @string/stop_service : @string/start_service}"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
